requirements.txt
requests
google-genai
python-dotenv
fastapi
uvicorn[standard]
apscheduler
jinja2
line-bot-sdk
psycopg2-binary
supabase
firebase-admin
pytest
httpx
